consolidate employee pages and remove duplicate queries
