+++ apps/member-profile/app/routes/_profile.companies_.$id._index.tsx
@@ -1,0 +1,22 @@
+import { type LoaderFunctionArgs, redirect } from '@remix-run/node';
+import { generatePath } from '@remix-run/react';
+
+import { getCompany } from '@oyster/core/employment/server';
+
+import { Route } from '@/shared/constants';
+
+export async function loader({ params }: LoaderFunctionArgs) {
+  const companyId = params.id as string;
+
+  const company = await getCompany({
+    include: ['reviews'],
+    select: [],
+    where: { id: companyId },
+  });
+
+  const to = Number(company?.reviews)
+    ? generatePath(Route['/companies/:id/reviews'], { id: companyId })
+    : generatePath(Route['/companies/:id/employees'], { id: companyId });
+
+  return redirect(to);
+}
