+++ apps/member-profile/app/shared/constants.ts
@@ -4,9 +4,8 @@ const ROUTES = [
   '/companies',
   '/companies/reviews/add',
   '/companies/:id',
+  '/companies/:id/employees',
   '/companies/:id/reviews',
-  '/companies/:id/current-employees',
-  '/companies/:id/past-employees',
   '/directory',
   '/directory/join',
   '/directory/join/1',
